# swagger_docs.yaml
openapi: 3.0.0
info:
  title: Assignment2 API
  description: API for managing books and fetching weather data.
  version: 1.0.0
servers:
  - url: http://localhost:8080

paths:
  /books:
    get:
      summary: Get all books
      operationId: getBooks
      responses:
        '200':
          description: Successfully retrieved books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
    post:
      summary: Add a new book
      operationId: postBooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: Successfully added book

  /book/{id}:
    put:
      summary: Update a book by ID
      operationId: updateBook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: Successfully updated book
        '404':
          description: Book not found
    delete:
      summary: Delete a book by ID
      operationId: deleteBook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successfully deleted book
        '404':
          description: Book not found

  /weather/{city}:
    get:
      summary: Get weather for a city
      operationId: getWeather
      parameters:
        - name: city
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved weather data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeatherResponse'
        '400':
          description: Invalid city name
        '404':
          description: City not found

components:
  schemas:
    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        year:
          type: integer
        genre:
          type: string

    WeatherResponse:
      type: object
      properties:
        city:
          type: string
        temperature:
          type: string
        condition:
          type: string
        feels_like:
          type: string
        humidity:
          type: string
        wind:
          type: string